separate Format functionality out of DateUtils into DateFormatter
