+++ src/main/java/org/javarosa/core/model/data/TimeData.java
@@ -17,6 +17,7 @@
 package org.javarosa.core.model.data;
 
 
+import org.javarosa.core.model.utils.DateFormatter;
 import org.javarosa.core.model.utils.DateUtils;
 import org.javarosa.core.util.externalizable.ExtUtil;
 import org.javarosa.core.util.externalizable.PrototypeFactory;
@@ -62,7 +63,7 @@ public class TimeData implements IAnswerData {
 
     @Override
     public String getDisplayText () {
-        return DateUtils.formatTime(d, DateUtils.FORMAT_HUMAN_READABLE_SHORT);
+        return DateFormatter.formatTime(d, DateFormatter.FORMAT_HUMAN_READABLE_SHORT);
     }
 
     @Override
@@ -77,7 +78,7 @@ public class TimeData implements IAnswerData {
 
     @Override
     public UncastData uncast() {
-        return new UncastData(DateUtils.formatTime(d, DateUtils.FORMAT_ISO8601));
+        return new UncastData(DateFormatter.formatTime(d, DateFormatter.FORMAT_ISO8601));
     }
 
     @Override
