+++ src/main/java/org/javarosa/core/model/utils/DateUtils.java
@@ -16,11 +16,9 @@
 
 package org.javarosa.core.model.utils;
 
-import org.javarosa.core.services.locale.Localization;
 import org.javarosa.core.util.MathUtils;
 import org.jetbrains.annotations.NotNull;
 import org.joda.time.LocalDateTime;
-import org.joda.time.format.DateTimeFormat;
 
 import java.time.LocalDate;
 import java.time.ZoneId;
@@ -52,25 +50,13 @@ public class DateUtils {
     }
 
 
-    public static final int MONTH_OFFSET = (1 - Calendar.JANUARY);
-
-    public static final int FORMAT_ISO8601 = 1;
-    public static final int FORMAT_HUMAN_READABLE_SHORT = 2;
-    public static final int FORMAT_HUMAN_READABLE_DAYS_FROM_TODAY = 5;
-    public static final int FORMAT_TIMESTAMP_SUFFIX = 7;
-
-    /**
-     * RFC 822
-     **/
-    public static final int FORMAT_TIMESTAMP_HTTP = 9;
-
     public static final long DAY_IN_MS = 86400000L;
 
     public DateUtils() {
         super();
     }
 
-    private static DateFields getFields(Date d) {
+    public static DateFields getFields(Date d) {
         return DateFields.getFields(d, null);
     }
 
@@ -83,204 +69,8 @@ public class DateUtils {
         return timezone == null ? ldt.toDate() : ldt.toDate(TimeZone.getTimeZone(timezone));
     }
 
-    private static LocalDateTime getLocalDateTime(DateFields f) {
+    public static LocalDateTime getLocalDateTime(DateFields f) {
         return new LocalDateTime(f.year, f.month, f.day, f.hour, f.minute, f.second, f.secTicks);
-    }
-
-    /* ==== FORMATTING DATES/TIMES TO STANDARD STRINGS ==== */
-
-    public static String formatDateTime(Date d, int format) {
-        if (d == null) {
-            return "";
-        }
-
-        DateFields fields = DateFields.getFields(d, format == FORMAT_TIMESTAMP_HTTP ? "UTC" : null);
-
-        String delim;
-        switch (format) {
-            case FORMAT_ISO8601:
-                delim = "T";
-                break;
-            case FORMAT_TIMESTAMP_SUFFIX:
-                delim = "";
-                break;
-            case FORMAT_TIMESTAMP_HTTP:
-                delim = " ";
-                break;
-            default:
-                delim = " ";
-                break;
-        }
-
-        return formatDate(fields, format) + delim + formatTime(fields, format);
-    }
-
-    public static String formatDate(Date d, int format) {
-        return (d == null ? "" : formatDate(DateFields.getFields(d, format == FORMAT_TIMESTAMP_HTTP ? "UTC" : null), format));
-    }
-
-    public static String formatTime(Date d, int format) {
-        return (d == null ? "" : formatTime(DateFields.getFields(d, format == FORMAT_TIMESTAMP_HTTP ? "UTC" : null), format));
-    }
-
-    private static String formatDate(DateFields f, int format) {
-        switch (format) {
-            case FORMAT_ISO8601:
-                return formatDateISO8601(f);
-            case FORMAT_HUMAN_READABLE_SHORT:
-                return formatDateColloquial(f);
-            case FORMAT_HUMAN_READABLE_DAYS_FROM_TODAY:
-                return formatDaysFromToday(f);
-            case FORMAT_TIMESTAMP_SUFFIX:
-                return formatDateSuffix(f);
-            case FORMAT_TIMESTAMP_HTTP:
-                return formatDateHttp(f);
-            default:
-                return null;
-        }
-    }
-
-    private static String formatTime(DateFields f, int format) {
-        switch (format) {
-            case FORMAT_ISO8601:
-                return formatTimeISO8601(f);
-            case FORMAT_HUMAN_READABLE_SHORT:
-                return formatTimeColloquial(f);
-            case FORMAT_TIMESTAMP_SUFFIX:
-                return formatTimeSuffix(f);
-            case FORMAT_TIMESTAMP_HTTP:
-                return formatTimeHttp(f);
-            default:
-                return null;
-        }
-    }
-
-    /**
-     * RFC 822
-     **/
-    private static String formatDateHttp(DateFields f) {
-        return format(f, "%a, %d %b %Y");
-    }
-
-    /**
-     * RFC 822
-     **/
-    private static String formatTimeHttp(DateFields f) {
-        return format(f, "%H:%M:%S GMT");
-    }
-
-    private static String formatDateISO8601(DateFields f) {
-        return f.year + "-" + intPad(f.month, 2) + "-" + intPad(f.day, 2);
-    }
-
-    private static String formatDateColloquial(DateFields f) {
-        String year = Integer.valueOf(f.year).toString();
-
-        //Normal Date
-        if (year.length() == 4) {
-            year = year.substring(2, 4);
-        }
-        //Otherwise we have an old or bizzarre date, don't try to do anything
-
-        return intPad(f.day, 2) + "/" + intPad(f.month, 2) + "/" + year;
-    }
-
-    private static String formatDateSuffix(DateFields f) {
-        return f.year + intPad(f.month, 2) + intPad(f.day, 2);
-    }
-
-    private static String formatTimeISO8601(DateFields f) {
-        String time = intPad(f.hour, 2) + ":" + intPad(f.minute, 2) + ":" + intPad(f.second, 2) + "." + intPad(f.secTicks, 3);
-
-        //Time Zone ops (1 in the first field corresponds to 'CE' ERA)
-        int milliday = ((f.hour * 60 + f.minute) * 60 + f.second) * 1000 + f.secTicks;
-        int offset = TimeZone.getDefault().getOffset(1, f.year, f.month - 1, f.day, f.dow, milliday);
-
-        //NOTE: offset is in millis
-        if (offset == 0) {
-            time += "Z";
-        } else {
-
-            //Start with sign
-            String offsetSign = offset > 0 ? "+" : "-";
-
-            int value = Math.abs(offset) / 1000 / 60;
-
-            String hrs = intPad(value / 60, 2);
-            String mins = ":" + intPad(value % 60, 2);
-
-            time += offsetSign + hrs + mins;
-        }
-        return time;
-    }
-
-    private static String formatTimeColloquial(DateFields f) {
-        return intPad(f.hour, 2) + ":" + intPad(f.minute, 2);
-    }
-
-    private static String formatTimeSuffix(DateFields f) {
-        return intPad(f.hour, 2) + intPad(f.minute, 2) + intPad(f.second, 2);
-    }
-
-    public static String format(Date d, String format) {
-        return format(getFields(d), format);
-    }
-
-    public static String format(DateFields f, String format) {
-        StringBuilder sb = new StringBuilder();
-
-        for (int i = 0; i < format.length(); i++) {
-            char c = format.charAt(i);
-
-            if (c == '%') {
-                i++;
-                if (i >= format.length()) {
-                    throw new RuntimeException("date format string ends with %");
-                } else {
-                    c = format.charAt(i);
-                }
-
-                if (c == '%') {            //literal '%'
-                    sb.append("%");
-                } else if (c == 'Y') {    //4-digit year
-                    sb.append(intPad(f.year, 4));
-                } else if (c == 'y') {    //2-digit year
-                    sb.append(intPad(f.year, 4).substring(2));
-                } else if (c == 'm') {    //0-padded month
-                    sb.append(intPad(f.month, 2));
-                } else if (c == 'n') {    //numeric month
-                    sb.append(f.month);
-                } else if (c == 'b') {    //short text month
-                    sb.append(getLocalDateTime(f).toString(DateTimeFormat.forPattern("MMM")));
-                } else if (c == 'd') {    //0-padded day of month
-                    sb.append(intPad(f.day, 2));
-                } else if (c == 'e') {    //day of month
-                    sb.append(f.day);
-                } else if (c == 'H') {    //0-padded hour (24-hr time)
-                    sb.append(intPad(f.hour, 2));
-                } else if (c == 'h') {    //hour (24-hr time)
-                    sb.append(f.hour);
-                } else if (c == 'M') {    //0-padded minute
-                    sb.append(intPad(f.minute, 2));
-                } else if (c == 'S') {    //0-padded second
-                    sb.append(intPad(f.second, 2));
-                } else if (c == '3') {    //0-padded millisecond ticks (000-999)
-                    sb.append(intPad(f.secTicks, 3));
-                } else if (c == 'a') {    //Three letter short text day
-                    sb.append(getLocalDateTime(f).toString(DateTimeFormat.forPattern("EEE")));
-                } else if (c == 'W') { // week of the year
-                    sb.append(f.week);
-                } else if (c == 'Z' || c == 'A' || c == 'B') {
-                    throw new RuntimeException("unsupported escape in date format string [%" + c + "]");
-                } else {
-                    throw new RuntimeException("unrecognized escape in date format string [%" + c + "]");
-                }
-            } else {
-                sb.append(c);
-            }
-        }
-
-        return sb.toString();
     }
 
     /* ==== PARSING DATES/TIMES FROM STANDARD STRINGS ==== */
@@ -493,7 +283,8 @@ public class DateUtils {
     public static Date roundDate(Date d) {
         if (d == null) return null;
         DateFields f = getFields(d);
-        return getDate(f.year, f.month, f.day);
+        return getDate(DateFields.of(f.year, f.month, f.day));
+//        return getDate(DateFields.of(d.getYear(), d.getMonth(), d.getDay()));
     }
 
     public static Date today() {
@@ -552,34 +343,6 @@ public class DateUtils {
 
 
     /* ==== Parsing to Human Text ==== */
-
-    /**
-     * Provides text representing a span of time.
-     *
-     * @param f The fields for the date to be compared against the current date.
-     * @return a string which is a human readable representation of the difference between
-     * the provided date and the current date.
-     */
-    private static String formatDaysFromToday(DateFields f) {
-        Date d = DateUtils.getDate(f);
-        int daysAgo = DateUtils.daysSinceEpoch(new Date()) - DateUtils.daysSinceEpoch(d);
-
-        if (daysAgo == 0) {
-            return Localization.get("date.today");
-        } else if (daysAgo == 1) {
-            return Localization.get("date.yesterday");
-        } else if (daysAgo == 2) {
-            return Localization.get("date.twoago", new String[]{String.valueOf(daysAgo)});
-        } else if (daysAgo > 2 && daysAgo <= 6) {
-            return Localization.get("date.nago", new String[]{String.valueOf(daysAgo)});
-        } else if (daysAgo == -1) {
-            return Localization.get("date.tomorrow");
-        } else if (daysAgo < -1 && daysAgo >= -6) {
-            return Localization.get("date.nfromnow", new String[]{String.valueOf(-daysAgo)});
-        } else {
-            return DateUtils.formatDate(f, DateUtils.FORMAT_HUMAN_READABLE_SHORT);
-        }
-    }
 
     /* ==== DATE OPERATIONS ==== */
 
@@ -695,9 +458,9 @@ public class DateUtils {
      * Tokenizes a string based on the given delimiter string
      *
      * @param str       The string to be split
-     * @param delimiter The delimeter to be used
+     * @param delimiter The delimiter to be used
      * @return An array of strings contained in original which were
-     * seperated by the delimeter
+     * seperated by the delimiter
      */
     public static List<String> split(String str, String delimiter, boolean combineMultipleDelimiters) {
 
@@ -722,18 +485,4 @@ public class DateUtils {
         return pieces;
     }
 
-    /**
-     * Converts an integer to a string, ensuring that the string
-     * contains a certain number of digits
-     *
-     * @param n   The integer to be converted
-     * @param pad The length of the string to be returned
-     * @return A string representing n, which has pad - #digits(n)
-     * 0's preceding the number.
-     */
-    public static String intPad(int n, int pad) {
-        String s = String.valueOf(n);
-        while (s.length() < pad) s = String.format("0%s", s);
-        return s;
-    }
 }
