+++ src/test/java/org/javarosa/core/model/utils/test/DateUtilsFormatTests.java
@@ -16,7 +16,7 @@
 package org.javarosa.core.model.utils.test;
 
 import org.javarosa.core.model.utils.DateFields;
-import org.javarosa.core.model.utils.DateUtils;
+import org.javarosa.core.model.utils.DateFormatter;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
@@ -25,19 +25,19 @@ public class DateUtilsFormatTests {
 
     @Test
     public void testFormatWeekOfYear() {
-        String week = DateUtils.format(DateFields.of(2018, 4, 1, 10, 20, 30, 400), "%W");
+        String week = DateFormatter.format(DateFields.of(2018, 4, 1, 10, 20, 30, 400), "%W");
         assertEquals("13", week);
 
-        week = DateUtils.format(DateFields.of(2018, 1, 1, 10, 20, 30, 400), "%W");
+        week = DateFormatter.format(DateFields.of(2018, 1, 1, 10, 20, 30, 400), "%W");
         assertEquals("1", week);
 
         // Week of year is based on what year the first thursday is. 1/1/2017 was a Sunday so 
         // it's actually the 52nd week of the previous year.
-        week = DateUtils.format(DateFields.of(2017, 1, 1, 10, 20, 30, 400), "%W");
+        week = DateFormatter.format(DateFields.of(2017, 1, 1, 10, 20, 30, 400), "%W");
         assertEquals("52", week);
 
         // 12/29/2020 is the 53rd week of the year 
-        week = DateUtils.format(DateFields.of(2020, 12, 29, 10, 20, 30, 400), "%W");
+        week = DateFormatter.format(DateFields.of(2020, 12, 29, 10, 20, 30, 400), "%W");
         assertEquals("53", week);
     }
 }
