+++ src/test/java/org/javarosa/core/model/utils/test/DateUtilsSCTOTests.java
@@ -1,6 +1,6 @@
 package org.javarosa.core.model.utils.test;
 
-import static org.javarosa.core.model.utils.DateUtils.FORMAT_ISO8601;
+import static org.javarosa.core.model.utils.DateFormatter.FORMAT_ISO8601;
 import static org.javarosa.test.utils.SystemHelper.withTimeZone;
 import static org.junit.Assert.assertEquals;
 
@@ -8,6 +8,8 @@ import java.util.Calendar;
 import java.util.Date;
 import java.util.SimpleTimeZone;
 import java.util.TimeZone;
+
+import org.javarosa.core.model.utils.DateFormatter;
 import org.javarosa.core.model.utils.DateUtils;
 import org.junit.Test;
 
@@ -17,7 +19,7 @@ public class DateUtilsSCTOTests {
     public void testParseDateTime() {
         withTimeZone(TimeZone.getTimeZone("GMT+02"), () -> {
             Date date = DateUtils.parseDateTime("2014-10-05T00:03:05.244+03");
-            String str = DateUtils.formatDateTime(date, FORMAT_ISO8601);
+            String str = DateFormatter.formatDateTime(date, FORMAT_ISO8601);
 
             assertEquals("2014-10-04T23:03:05.244+02:00", str);
         });
@@ -27,7 +29,7 @@ public class DateUtilsSCTOTests {
     public void testParseDateTime_withDST() {
         withTimeZone(buildDstTimeZone(), () -> {
             Date date = DateUtils.parseDateTime("2014-10-05T00:03:05.244+03");
-            String str = DateUtils.formatDateTime(date, FORMAT_ISO8601);
+            String str = DateFormatter.formatDateTime(date, FORMAT_ISO8601);
 
             assertEquals("2014-10-05T00:03:05.244+03:00", str);
         });
@@ -38,7 +40,7 @@ public class DateUtilsSCTOTests {
         withTimeZone(TimeZone.getTimeZone("GMT+02"), () -> {
             String time = "12:03:05.011+03";
             Date date = DateUtils.parseTime(time);
-            String formatted = DateUtils.formatTime(date, FORMAT_ISO8601);
+            String formatted = DateFormatter.formatTime(date, FORMAT_ISO8601);
             assertEquals("11:03:05.011+02:00", formatted);
         });
     }
